+++ Engine/Source/Runtime/Core/Public/ProfilingDebugging/FormatArgsTrace.h
@@ -121,7 +121,7 @@ private:
 		*TypeCodesPtr++ = FormatArgTypeCode_CategoryString | sizeof(CharType);
 		if (Argument != nullptr)
 		{
-			uint16 Length = (TCString<CharType>::Strlen(Argument) + 1) * sizeof(CharType);
+			uint16 Length = (uint16)((TCString<CharType>::Strlen(Argument) + 1) * sizeof(CharType));
 			memcpy(PayloadPtr, Argument, Length);
 			PayloadPtr += Length;
 		}
